Add Kani physics harnesses, strict-verify CI, and clippy cleanup
Port the 17 Part 1 Kani proof harnesses into src/verification/physics.rs
(classical mechanics, gravitation, Mat3::inverse, bisection, statistics,
entropy, angle normalization), taking the crate from 3 harnesses to 20.
These prove panic-freedom and range properties over the whole f64 domain
of each stated precondition, which is the right tool for the defensive
guards that unit tests cannot reach by construction.

Add .github/workflows/verify.yml with property, Kani, Miri and strict
clippy jobs. Two deliberate departures from a naive strict setup: the
float-accuracy lints stay at warn level, because suboptimal_flops alone
fires ~2800 times and rewriting those as mul_add changes rounding, so
each needs review rather than blanket denial; and Miri is scoped to the
core numerics modules, because the crate contains no unsafe and
interpreting the whole suite would dominate CI.

Clean every default clippy lint outside the modules still being worked
on. Substantive fixes rather than suppressions: replace 11 hardcoded
constant approximations with std::f64::consts (tightening the assertions,
since the true values are exactly those constants), correct 16 float
literals carrying more precision than f64 holds or grouped
inconsistently, take a slice instead of &mut Vec in joseph_update, drop
a dead sin_cos computation in the Vincenty solver whose result was
overwritten before use, replace a manual slice copy and a repeated-push
loop in the spline and knot code, and turn vector_calculus's file-level
/// comments into //! module docs.

Four allowances are deliberate and documented at the crate root: index
loops that mirror matrix notation, frozen roadmap signatures with more
than seven arguments, !(x > 0.0) guards that also reject NaN, and the
inherent next() methods on the quasi-random sequences, which cannot be
Iterator because they borrow self mutably and yield a fresh Vec.

Also add brute-force and identity tests for previously unexercised
spatial and linalg surface: quadtree and kd-tree queries checked against
exhaustive enumeration, SDF primitives checked against closed forms and
degenerate-case identities, Rect and Polyline bounds, Homography
composition, Mat4 trace/inverse/Mul, and Mat3::principal_axes.

+++ src/radiation.rs
@@ -86,7 +86,7 @@ pub fn view_factor_parallel_plates(width: f64, height: f64, separation: f64) -> 
     let b = 1.0 + y * y;
     let c = 1.0 + x2 + y * y;
 
-    let term_ln = ((a * b / c) as f64).sqrt().ln();
+    let term_ln = (a * b / c).sqrt().ln();
     let term_x = x * b.sqrt() * (x / b.sqrt()).atan() - x * x.atan();
     let term_y = y * a.sqrt() * (y / a.sqrt()).atan() - y * y.atan();
